Create Top 10 Cities in Top 10 Countries
For this query, I wanted to find the number of customers in each country and rank them most to least. First, I asked to return the “country” column from the “country” table. Then, I asked to return a COUNT of the values in the “customer_id” column of the “customer” table under the alias “number_of_customers.” To join the two tables, I asked to match “address_id” from the “customer” and “address” tables, then match “city_id” from the “address” and “city” tables, then match “country_id” on the “city” and “country” tables. I then asked to GROUP the results by “country”, in the order of “number_of_customers”, descending (from most to least), with a LIMIT of 10.
This query was very similar to the first, with the addition of filtering the results WHERE the “country” ‘IN’cludes the information based on the cities with the highest number of customers. Basically, reentering the same query to filter again.